main: Use new power pin, parse CID properly
This patch passes the new power pin to the SD init function.

Additionally, we now properly parse the CID that gets returned.  We
do this by stripping away the UNSTUFF_BITS macro that we previously
stole from the Linux kernel, and manually unpacking the bits ourselves.

Granted this won't work so well on systems that are byteswapped, but
it should serve for the time being.