spicode: Get compiling under -pedantic -Werror
This just seems like a good idea.  It's an exercise in doing things
correctly, at least.  In doing so, we've been forced to specify C99,
and this has removed some functions such as usleep().  So far we've
implemented hokey replacements, but overall this will be a good thing.

Note that compiling under ARM hasn't yet been tested.  As it's a
different platform, it's likely to not work.  I'll fix things once
I get to an ARM system.