spicode: Get rid of data/command distinction
Since we're using TCP for both the command and data pipes, and we
have a well-defined packet structure, the split between data and
command doesn't make as much sense anymore.  Therefore, we remove
the command pipe and instead start reading all data over the data
pipe.

Note that this means the program is no longer able to be controlled
over a telnet session.  You will need a special client in order to
drive this program now.